Add an rsync file staging provider
this should support cloud-based execution

so rsync needs to run co-located with the task

it should assume submit side is running some kind of rsync-compatible server
that can be passed to rsync commandline (which by default then is rsync+ssh)

authentication is awkward: how can the remote side authenticate back without
ridiculous security requirements?

to begin with, perhaps specify an ssh private key that will be moved around
the network to worker nodes, ideally which only has rsync access to the user
account. (but ... ick ... but ssh doesn't really have delegation)

at the very beginning for my prototyping locally, there will be an ssh agent
most likely available because I'll be running tasks where that exists. In
which case I can treat auth as being completely external. It would also be
an option to say: the remote account needs to have auth configured via
other options, not by the provider: for example, worker init should set this
up appropriately, so that it becomes a more complex config setup rather than
part of the rsync code. That is probably a reasonable approach to take for
an initial implementation.


Also contains some bugfixes which should move into separate PR