Fix a dependency bug that only shows with certain stageout patterns.
When a staging provider wishes to edit its File object in stage_in
(for example, to set the local file path) but does not wish to insert
anything into the task graph (for example, because staging will
happen in a task wrapper), and when that File object was copied from
a DataFuture that came from a previous task's output list, then
either:

 * the stageout provider could return None to keep the dependency
   on the earlier stageout in place
or

 * it could return the modified File object so that the stagein wrapper
   sees the modifications.

but it could not do both.

This patch changes the File copying mechanism so that returning None
will both keep the dependency on the earlier stageout, and pass on
the modified file rather than the original application-level File.

It does this by creating a new DataFuture which completes when the
original DataFuture completes, but contains the new copy of the
File rather than the original File.

The bug fixed here was introduced as part of PR#1272.
I discovered it while developing a provider which would do file: scheme
staging in task wrappers, as described above, which is not currently
tested in the test suite.